Break up Brief + Document
