+++ Sources/Packages/Sources/SecureEnclaveSecretKit/Documentation.docc/Documentation.md
@@ -1,3 +1,3 @@
-# ````SecureEnclaveSecretKit````
+# ``SecureEnclaveSecretKit``
 
 SecureEnclaveSecretKit contains implementations of SecretKit protocols backed by the Secure Enclave.
