+++ Sources/Packages/Sources/SmartCardSecretKit/Documentation.docc/Documentation.md
@@ -1,3 +1,3 @@
-# ````SmartCardSecretKit````
+# ``SmartCardSecretKit``
 
 SmartCardSecretKit contains implementations of SecretKit protocols backed by a Smart Card.
